+++ concurrency/src/mingwX64Main/kotlin/pw/binom/concurrency/Worker.kt
@@ -3,5 +3,5 @@ package pw.binom.concurrency
 import platform.windows.Sleep
 
 actual fun Worker.Companion.sleep(deley: Long){
-    Sleep(deley.toUInt())
+    platform.windows.Sleep(deley.toUInt())
 }