+++ history.md
@@ -2,6 +2,8 @@
 
 ### 0.1.20
 * Added charset support
+* Added BytesBitset
+* Added WorkerPool
 
 ### 0.1.19
-* TarantoolDB Connector 
+* TarantoolDB Connector